+++ config/test.exs
@@ -61,8 +61,6 @@ config :web_push_encryption, :vapid_details,
 
 config :web_push_encryption, :http_client, Pleroma.Web.WebPushHttpClientMock
 
-config :pleroma_job_queue, disabled: true
-
 config :pleroma, Oban,
   queues: false,
   prune: :disabled